fix(episodes_edit.php): ensure proper generation of missing GUID value
Fix up the code for round-tripping or generating the GUID value so that we don't get an incorrect
file path in the GUID. To simplify the round-tripping, we'll also skip generating a missing GUID
when building the edit form since that wasn't actually necessary. Just generate once at save if
there's no GUID already available for the episode.